refactor : 패키지 구조를 좀더 알아보기 쉽게 변경하였습니다.
