+++ src/main/java/org/prgrms/kdtspringdemo/io/console/Input.java
@@ -1,4 +1,4 @@
-package org.prgrms.kdtspringdemo.io;
+package org.prgrms.kdtspringdemo.io.console;
 
 public interface Input {
     String getInput();
