+++ src/main/java/org/prgrms/kdtspringdemo/io/file/CSVReader.java
@@ -1,0 +1,4 @@
+package org.prgrms.kdtspringdemo.io.file;
+
+public class CSVReader {
+}
