+++ src/main/java/org/prgrms/kdtspringdemo/voucher/repository/MemoryVoucherRepository.java
@@ -1,5 +1,6 @@
-package org.prgrms.kdtspringdemo.voucher;
+package org.prgrms.kdtspringdemo.voucher.repository;
 
+import org.prgrms.kdtspringdemo.voucher.model.Voucher;
 import org.springframework.beans.factory.DisposableBean;
 import org.springframework.beans.factory.InitializingBean;
 import org.springframework.context.annotation.Profile;
