+++ src/main/java/org/prgrms/kdtspringdemo/voucher/repository/VoucherRepository.java
@@ -1,4 +1,6 @@
-package org.prgrms.kdtspringdemo.voucher;
+package org.prgrms.kdtspringdemo.voucher.repository;
+
+import org.prgrms.kdtspringdemo.voucher.model.Voucher;
 
 import java.util.List;
 import java.util.Optional;
